// src/model/adapters/musicAssistant/types/musicAssistantTypes.ts
export enum RepeatMode {
  OFF = 'off', // no repeat at all
  ONE = 'one', // repeat current/single track
  ALL = 'all', // repeat entire queue
}

export enum PlaybackState {
  IDLE = 'idle',
  PAUSED = 'paused',
  PLAYING = 'playing',
}
// ...
export interface MediaItemImage {
  type: ImageType;
  path: string;
  provider: string;
  remotely_accessible: boolean;
}
// ...
export interface Artist extends MediaItem {}

export interface Album extends MediaItem {
  year?: number;
  artists: Array<ItemMapping | Artist>;
  album_type: AlbumType;
}

export interface Track extends MediaItem {
  duration: number;
  artists: Array<ItemMapping | Artist>;
  // album track only
  album: ItemMapping | Album;
  disc_number?: number;
  track_number?: number;
}

export interface Playlist extends MediaItem {
  owner: string;
  is_editable: boolean;
}

export interface Radio extends MediaItem {}

export interface Audiobook extends MediaItem {
  publisher: string;
  authors: string[];
  narrators: string[];
  duration: number;
  fully_played?: boolean;
  resume_position_ms?: number;
}

export interface Podcast extends MediaItem {
  publisher?: string;
  total_episodes?: number;
}

export interface PodcastEpisode extends MediaItem {
  position: number;
  podcast: Podcast | ItemMapping;
  duration: number;
  fully_played?: boolean;
  resume_position_ms?: number;
}

export interface BrowseFolder extends MediaItem {
  path?: string;
  image?: MediaItemImage;
}
export interface RecommendationFolder extends BrowseFolder {
  icon?: string;
  items: MediaItemTypeOrItemMapping[];
}

export type MediaItemType =
  | Artist
  | Album
  | Track
  | Radio
  | Playlist
  | Audiobook
  | Podcast
  | PodcastEpisode
  | BrowseFolder;

export type PlayableMediaItemType = Track | Radio | Audiobook | PodcastEpisode;
// ...
export interface StreamDetails {
  provider: string;
  item_id: string;
  audio_format: AudioFormat;
  media_type: MediaType;
  stream_metadata?: StreamMetadata;
  duration?: number;

  queue_id?: string;
  fade_in?: boolean;
  loudness?: number;
  loudness_album?: number;
  prefer_album_loudness?: boolean;
  target_loudness?: number;
  volume_normalization_mode?: VolumeNormalizationMode;
  volume_normalization_gain_correct?: number;
  // This contains the DSPDetails of all players in the group.
  // In case of single player playback, dict will contain only one entry.
  dsp?: Record<string, DSPDetails>;
}

// queue_item

export interface QueueItem {
  queue_id: string;
  queue_item_id: string;
  name: string;
  duration: number;
  sort_index: number;
  streamdetails?: StreamDetails;
  media_item?: PlayableMediaItemType;
  image?: MediaItemImage;
  available: boolean;
}

// player_queue

export interface PlayerQueue {
  queue_id: string;
  active: boolean;
  display_name: string;
  available: boolean;
  items: number;
  shuffle_enabled: boolean;
  dont_stop_the_music_enabled: boolean;
  repeat_mode: RepeatMode;
  current_index?: number;
  index_in_buffer?: number;
  elapsed_time: number;
  /**
   * UTC timestamp (seconds since epoch) when `elapsed_time` was last updated.
   *
   * Semantics/units:
   * - `elapsed_time` is expressed in seconds (number, can be fractional).
   * - `elapsed_time_last_updated` is a UTC timestamp in seconds since epoch.
   *   Convert to milliseconds (multiply by 1000) when comparing to Date.now().
   *
   * Use this timestamp to compute the current progress while playback is
   * ongoing by adding (now - elapsed_time_last_updated*1000)/1000 to
   * `elapsed_time`.
   */
  elapsed_time_last_updated: number;
  state: PlaybackState;
  current_item?: QueueItem;
  next_item?: QueueItem;
  radio_source: MediaItemType[];
}